Prep lesson for making image optional
