+++ src/components/AddItem.js
@@ -4,6 +4,7 @@ import {UIImagePickerManager} from 'NativeModules';
 import {setNewItemImage, setNewItemTitle, commitNewItem, clearPendingNewItem} from '../actionCreators/stuff';
 import {MKColor, MKButton, MKTextField} from 'react-native-material-kit';
 import {addItemStyles, floatingButton} from '../styles';
+import SelectImage from './SelectImage';
 
 @connect(state => {
   return {
@@ -16,32 +17,6 @@ import {addItemStyles, floatingButton} from '../styles';
     this.state = {
       keyboardSpace: 0,
     };
-  }
-
-  addImage() {
-    const options = {
-      maxWidth: 400,
-      maxHeight: 400,
-      quality: 0.9,
-      allowsEditing: true,
-      noData: false,
-      storageOptions: {
-        skipBackup: true,
-        path: 'images',
-      },
-    };
-
-    // TODO this might be eating too much space :-/
-
-    UIImagePickerManager.showImagePicker(options, (didCancel, response) => {
-      if (!didCancel) if (!response.customButton) {
-        // You can display the image using either:
-        //const source = {uri: 'data:image/jpeg;base64,' + response.data, isStatic: true};
-        const source = {uri: response.uri.replace('file://', ''), isStatic: true};
-
-        this.props.dispatch(setNewItemImage(source.uri));
-      }
-    });
   }
 
   commit() {
@@ -61,6 +36,7 @@ import {addItemStyles, floatingButton} from '../styles';
   }
 
   render() {
+    // LESSON Change instruction text and include SelectImage below it.
     return <View style={styles.scene}>
       <ScrollView
         ref='scroller'
@@ -74,16 +50,9 @@ import {addItemStyles, floatingButton} from '../styles';
         onKeyboardDidHide={(e) => this.setState({keyboardSpace: 0})}
       >
         <View style={styles.innerScroller}>
-          <Text style={styles.instructions}>Tap below to add a picture, then name your item.</Text>
-          <TouchableOpacity
-            onPress={() => this.addImage()}
-            style={floatingButton}
-          >
-            <Image
-              style={styles.previewImage}
-              source={{uri: this.props.previewImageUri, isStatic: true}}
-            />
-          </TouchableOpacity>
+
+          <Text style={styles.instructions}>Name your item below.</Text>
+
           <MKTextField
             ref='nameInput'
             style={styles.nameInput}
